requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
executing==1.2.0
idna==3.4
inflection==0.5.1
ipdb==0.13.13
ipython==8.12.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
openapi-codec==1.3.2
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.4
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
